requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
colorama==0.4.6
wheel